tugraz: import booster rules for J&J
This imports the booster rule for the J&J vaccine from the official rule set.
Change validity from 360 to 270 according to the TU Graz rules.

This makes HCERTs for a second dose of J&J valid immediately. Before this one still
had to wait 22 days even after the second dose.