tugraz: remove all time based restrictions
Every type of valid HCERT is now valid, so this comes down to
3G without any time based limits.

We still check:

* generic checks for valid HCERTs
* if the used RAT test product is OK and if it is negative
* if this is for covid-19
* if the vaccine is OK
* if it's on the global "block list"

Both the trust anchor and the HCERT signature still have an expiration date,
so there is still a time limit besides these rules. But this is enforced by
the application on the client device.